gh-119659: Move `@no_rerun` to `test.support`
